shared privates weak map
